+++ src/components/ui/alert.jsx
@@ -1,22 +1,26 @@
 import React from 'react'
+import { cn } from '@/lib/utils'
 
-const Alert = ({ children, className = '', ...props }) => {
-  return (
-    <div
-      className={`relative w-full rounded-lg border p-4 [&>svg~*]:pl-7 [&>svg+div]:translate-y-[-3px] [&>svg]:absolute [&>svg]:left-4 [&>svg]:top-4 [&>svg]:text-foreground ${className}`}
-      {...props}
-    >
-      {children}
-    </div>
-  )
-}
+const Alert = React.forwardRef(({ className, ...props }, ref) => (
+  <div
+    ref={ref}
+    role="alert"
+    className={cn(
+      'relative w-full rounded-lg border p-4 [&>svg~*]:pl-7 [&>svg+div]:translate-y-[-3px] [&>svg]:absolute [&>svg]:left-4 [&>svg]:top-4 [&>svg]:text-foreground',
+      className
+    )}
+    {...props}
+  />
+))
+Alert.displayName = 'Alert'
 
-const AlertDescription = ({ children, className = '', ...props }) => {
-  return (
-    <div className={`text-sm [&_p]:leading-relaxed ${className}`} {...props}>
-      {children}
-    </div>
-  )
-}
+const AlertDescription = React.forwardRef(({ className, ...props }, ref) => (
+  <div
+    ref={ref}
+    className={cn('text-sm [&_p]:leading-relaxed', className)}
+    {...props}
+  />
+))
+AlertDescription.displayName = 'AlertDescription'
 
 export { Alert, AlertDescription }
