+++ src/components/ui/badge.jsx
@@ -1,18 +1,28 @@
 import React from 'react'
 
-const Badge = ({ children, variant = 'default', className = '', ...props }) => {
-  const variants = {
-    default: 'bg-primary text-primary-foreground hover:bg-primary/80',
-    secondary: 'bg-secondary text-secondary-foreground hover:bg-secondary/80',
-    destructive: 'bg-destructive text-destructive-foreground hover:bg-destructive/90',
-    outline: 'text-foreground border border-input hover:bg-accent hover:text-accent-foreground'
-  }
+// Simplified version without cn dependency
+const Badge = React.forwardRef(({ className = '', variant = 'default', ...props }, ref) => {
+  const baseClasses = 'inline-flex items-center rounded-full border px-2.5 py-0.5 text-xs font-semibold transition-colors'
   
+  const variantClasses = {
+    default: 'border-transparent bg-gray-100 text-gray-800',
+    secondary: 'border-transparent bg-blue-100 text-blue-800', 
+    destructive: 'border-transparent bg-red-100 text-red-800',
+    outline: 'border-gray-300 text-gray-700',
+    success: 'border-transparent bg-green-100 text-green-800 border-green-300'
+  }
+
+  const combinedClass = `${baseClasses} ${variantClasses[variant] || variantClasses.default} ${className}`
+
   return (
-    <span className={`inline-flex items-center rounded-full border px-2.5 py-0.5 text-xs font-semibold transition-colors focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2 ${variants[variant]} ${className}`} {...props}>
-      {children}
-    </span>
+    <span
+      ref={ref}
+      className={combinedClass}
+      {...props}
+    />
   )
-}
+})
+
+Badge.displayName = 'Badge'
 
 export { Badge }
