+++ src/components/ui/button.jsx
@@ -1,27 +1,31 @@
 import React from 'react'
+import { cn } from '@/lib/utils'
 
-const Button = React.forwardRef(({ className = '', variant = 'default', size = 'default', ...props }, ref) => {
-  const baseClasses = 'inline-flex items-center justify-center rounded-md text-sm font-medium transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:opacity-50 disabled:pointer-events-none ring-offset-background'
-  
-  const variants = {
-    default: 'bg-primary text-primary-foreground hover:bg-primary/90',
-    destructive: 'bg-destructive text-destructive-foreground hover:bg-destructive/90',
-    outline: 'border border-input hover:bg-accent hover:text-accent-foreground',
-    secondary: 'bg-secondary text-secondary-foreground hover:bg-secondary/80',
-    ghost: 'hover:bg-accent hover:text-accent-foreground',
-    link: 'underline-offset-4 hover:underline text-primary'
-  }
-  
-  const sizes = {
-    default: 'h-10 py-2 px-4',
-    sm: 'h-9 px-3 rounded-md',
-    lg: 'h-11 px-8 rounded-md',
-    icon: 'h-10 w-10'
-  }
-  
-  const classes = `${baseClasses} ${variants[variant]} ${sizes[size]} ${className}`
-  
-  return <button className={classes} ref={ref} {...props} />
+const Button = React.forwardRef(({ className, variant = 'default', size = 'default', ...props }, ref) => {
+  return (
+    <button
+      className={cn(
+        'inline-flex items-center justify-center whitespace-nowrap rounded-md text-sm font-medium ring-offset-background transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50',
+        {
+          'bg-primary text-primary-foreground hover:bg-primary/90': variant === 'default',
+          'border border-input bg-background hover:bg-accent hover:text-accent-foreground': variant === 'outline',
+          'bg-destructive text-destructive-foreground hover:bg-destructive/90': variant === 'destructive',
+          'bg-secondary text-secondary-foreground hover:bg-secondary/80': variant === 'secondary',
+          'hover:bg-accent hover:text-accent-foreground': variant === 'ghost',
+          'bg-background hover:bg-accent hover:text-accent-foreground': variant === 'link',
+        }[variant],
+        {
+          'h-10 px-4 py-2': size === 'default',
+          'h-9 rounded-md px-3': size === 'sm',
+          'h-11 rounded-md px-8': size === 'lg',
+          'h-12 rounded-md px-3': size === 'icon',
+        }[size],
+        className
+      )}
+      ref={ref}
+      {...props}
+    />
+  )
 })
 
 Button.displayName = 'Button'
