+++ src/components/ui/progress.jsx
@@ -1,14 +1,21 @@
 import React from 'react'
+import { cn } from '@/lib/utils'
 
-const Progress = ({ value = 0, className = '', ...props }) => {
-  return (
-    <div className={`relative h-4 w-full overflow-hidden rounded-full bg-secondary ${className}`} {...props}>
-      <div
-        className="h-full w-full flex-1 bg-primary transition-all"
-        style={{ transform: `translateX(-${100 - (value || 0)}%)` }}
-      />
-    </div>
-  )
-}
+const Progress = React.forwardRef(({ value, className, ...props }, ref) => (
+  <div
+    ref={ref}
+    className={cn(
+      'relative h-4 w-full overflow-hidden rounded-full bg-secondary',
+      className
+    )}
+    {...props}
+  >
+    <div
+      className="h-full w-full flex-1 bg-primary transition-all"
+      style={{ transform: `translateX(-${100 - (value || 0)}%)` }}
+    />
+  </div>
+))
+Progress.displayName = 'Progress'
 
 export { Progress }
