+++ src/components/ui/tabs.jsx
@@ -1,55 +1,70 @@
-import React, { useState } from 'react'
+import React, { createContext, useContext, useState } from 'react'
+import { cn } from '@/lib/utils'
 
-const TabsContext = React.createContext()
+const TabsContext = createContext({})
 
-const Tabs = ({ children, defaultValue, className = '', ...props }) => {
+export const Tabs = ({ defaultValue, className, children, ...props }) => {
   const [value, setValue] = useState(defaultValue)
-  
+
   return (
     <TabsContext.Provider value={{ value, setValue }}>
-      <div className={className} {...props}>
+      <div className={cn('flex flex-col', className)} {...props}>
         {children}
       </div>
     </TabsContext.Provider>
   )
 }
 
-const TabsList = ({ children, className = '', ...props }) => {
-  return (
-    <div className={`inline-flex h-10 items-center justify-center rounded-md bg-muted p-1 text-muted-foreground ${className}`} {...props}>
-      {children}
-    </div>
-  )
-}
+export const TabsList = React.forwardRef(({ className, ...props }, ref) => (
+  <div
+    ref={ref}
+    className={cn(
+      'inline-flex h-10 items-center justify-center rounded-md bg-muted p-1 text-muted-foreground',
+      className
+    )}
+    {...props}
+  />
+))
+TabsList.displayName = 'TabsList'
 
-const TabsTrigger = ({ children, value, className = '', ...props }) => {
-  const { value: selectedValue, setValue } = React.useContext(TabsContext)
+export const TabsTrigger = React.forwardRef(({ children, value, className, ...props }, ref) => {
+  const { value: selectedValue, setValue } = useContext(TabsContext)
   
   return (
     <button
-      className={`inline-flex items-center justify-center whitespace-nowrap rounded-sm px-3 py-1.5 text-sm font-medium ring-offset-background transition-all focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50 ${
-        selectedValue === value
-          ? 'bg-background text-foreground shadow-sm'
-          : 'hover:bg-muted hover:text-foreground'
-      } ${className}`}
+      ref={ref}
+      className={cn(
+        'inline-flex items-center justify-center whitespace-nowrap rounded-sm px-3 py-1.5 text-sm font-medium ring-offset-background transition-all focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50 cursor-pointer',
+        selectedValue === value 
+          ? 'bg-background text-foreground shadow-sm' 
+          : 'hover:bg-background/50 hover:text-foreground',
+        className
+      )}
       onClick={() => setValue(value)}
       {...props}
     >
       {children}
     </button>
   )
-}
+})
+TabsTrigger.displayName = 'TabsTrigger'
 
-const TabsContent = ({ children, value, className = '', ...props }) => {
-  const { value: selectedValue } = React.useContext(TabsContext)
+export const TabsContent = React.forwardRef(({ value, className, ...props }, ref) => {
+  const { value: selectedValue } = useContext(TabsContext)
   
-  if (selectedValue !== value) return null
+  if (selectedValue !== value) {
+    return null
+  }
   
   return (
-    <div className={`mt-2 ring-offset-background focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 ${className}`} {...props}>
-      {children}
-    </div>
+    <div
+      ref={ref}
+      className={cn(
+        'mt-2 ring-offset-background focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2',
+        className
+      )}
+      {...props}
+    />
   )
-}
-
-export { Tabs, TabsList, TabsTrigger, TabsContent }
+})
+TabsContent.displayName = 'TabsContent'
