+++ src/hooks/useDarkMode.js
@@ -1,0 +1,17 @@
+import { useState, useEffect } from 'react'
+
+export const useDarkMode = () => {
+  const [darkMode, setDarkMode] = useState(false)
+
+  useEffect(() => {
+    if (darkMode) {
+      document.documentElement.classList.add('dark')
+    } else {
+      document.documentElement.classList.remove('dark')
+    }
+  }, [darkMode])
+
+  const toggleDarkMode = () => setDarkMode(!darkMode)
+
+  return { darkMode, toggleDarkMode }
+}
