+++ src/test/setup.js
@@ -1,0 +1,1 @@
+import '@testing-library/jest-dom'
